fix comment ( xóa spacing, add MARK, chuyển ext )
