+++ FinalProject/Module/1. Login/LoginViewModel.swift
@@ -10,13 +10,15 @@ import Foundation
 import UIKit
 
 final class LoginViewModel {
-    
+    // MARK: Properties
     let networkManager: NetworkManager
-
+    
+    // MARK: Init
     init(networkManager: NetworkManager = NetworkManager.shared) {
         self.networkManager = networkManager
     }
-
+    
+    // MARK: Function
     func login(phone: String, pw: String, completion: @escaping APICompletion ) {
         networkManager.login(phone: phone, pw: pw) { result in
             switch result {
